requirements.txt
tensorflow==2.10.0
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.2.0
matplotlib==3.6.0
tkinter==0.1.0  # Install using the system package manager if necessary
scikit-learn==1.1.2
h5py==3.7.0
